[CI] Migrate LoRA tests to test/registered/lora/
Part of CI reorganization effort (#13808).

Changes:
- Move lora_utils.py to python/sglang/test/lora_utils.py
- Migrate 6 LoRA test files from test/srt/lora/ to test/registered/lora/
- Add CI registration with appropriate suites:
  - 5 tests -> stage-b-test-small-1-gpu
  - 1 test (TP) -> stage-b-test-small-2-gpu
- Add stage-b-test-small-2-gpu to per-commit CUDA suites
- Remove LoRA entries from old test/srt/run_suite.py
- Clean up sys.path hacks, use proper sglang.test imports

Not migrated (nightly test):
- test_lora_hf_sgl_logprob_diff.py (already registered)
