+++ test/registered/lora/test_lora_eviction.py
@@ -19,8 +19,11 @@ from typing import Dict, List, Tuple
 
 import torch
 
+from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.runners import SRTRunner
 from sglang.test.test_utils import CustomTestCase
+
+register_cuda_ci(est_time=224, suite="stage-b-test-small-1-gpu")
 
 PROMPTS = [
     "AI is a field of computer science focused on",
