+++ test/registered/lora/test_lora_update.py
@@ -23,6 +23,7 @@ import requests
 import torch
 
 from sglang.srt.utils import kill_process_tree
+from sglang.test.ci.ci_register import register_cuda_ci
 from sglang.test.runners import SRTRunner
 from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
@@ -32,6 +33,8 @@ from sglang.test.test_utils import (
     is_in_ci,
     popen_launch_server,
 )
+
+register_cuda_ci(est_time=451, suite="stage-b-test-small-1-gpu")
 
 PROMPTS = [
     "SGL is a",
